src/backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.4.0
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.0
requests==2.31.0